Build in a very basic "tool.spin-test" config
This allows specifying in the Spin.toml where the test lives and how to
build it. `spin-test` will automatically find and build the test before
running it.

This does place a limit on only one test component per app component.
This is probably fine for now, but we might want to adjust this soon.

Signed-off-by: Ryan Levick <ryan.levick@fermyon.com>