Add default headers to request in router
Spin runtimes are expected to add certain headers to each request as can be seen [here](https://github.com/fermyon/spin/blob/efac5cf2249987b80bcb2c11a76564547df72327/crates/trigger-http/src/lib.rs#L498-L534). The most appropriate place for this to happen in `spin-test` is in the router. However, `wasi:http/types/incoming-request` does not allow for modifying the headers as the headers are set to immutable (and any mutating function called on them returns an `wasi:http/types/header-error.immutable`). This is a rather fundamental shortcoming of `wasi:http@0.2.0` that we won't be able to get around for the foreseeable future.

To work around the above, we add an optional `incoming-body` to `fermyon:spin-test/http-helper/new-request` which allows the non-body parts of the request to come from the `outgoing-request` while the body comes from the `incoming-request`.

Signed-off-by: Ryan Levick <ryan.levick@fermyon.com>